Support sending modifiers immediately with double click
Previously, the buttons in the web UI allowed the user to simulate holding a modifier key on the next keypress, but there was no way to send, for example, the OS keypress on its own. This change allows users to double-click the modifier key buttons to simulate a keypress of that particular key.

Fixes #37