Add keyup event to on-screen-keyboard
app.js needs to keep track of which modifiers are pressed and unpressed, so the on-screen-keyboard needs to also emit events indicating when a key has been released.

This fixes a regression in which holding down the shift key began sending multiple shift key press events to the backend, when it should only send a single event.