Support altGraph key on on-screen keyboard
Previously, the on-screen keyboard never distinguished between left or right alt keys being pressed, but on certain keyboard layouts, the right alt is treated as AltGraph and has special meaning.

This change causes pressing the right alt key on the on-screen keyboard to set altGraphKey: true in the socket.io message to the backend, so that the backend translates it properly as a right alt modifier.

The behavior is not perfect in that it will treat holding the right alt key the same as holding both alt keys at once, but hopefully this moves us closer to a complete solution.

Fixes #350