Convert from keyCode to key
TinyPilot currently uses the keyCode field from the JavaScript keyboard event to communicate which key was pressed. The keyCode field is deprecated in favor of the code field.

The code field has values like KeyA which don't necessarily match the labels on the keyboard for non-US keyboards. For example, on US QWERTY keyboards, pressing the key in the upper left corner of the letters area (Q) sends KeyQ. Pressing the same key on a French keyboard, where the target system is configured for French AZERTY layout sends KeyQ even though the key is labeled A on a French keyboard.

By switching from keyCode to code, TinyPilot should be able to escape the work of creating manual mappings for every keyboard layout.

Fixes #354