Convert from keyCode to key (#355)
* Convert from keyCode to key

TinyPilot currently uses the keyCode field from the JavaScript keyboard event to communicate which key was pressed. The keyCode field is deprecated in favor of the code field.

The code field has values like KeyA which don't necessarily match the labels on the keyboard for non-US keyboards. For example, on US QWERTY keyboards, pressing the key in the upper left corner of the letters area (Q) sends KeyQ. Pressing the same key on a French keyboard, where the target system is configured for French AZERTY layout sends KeyQ even though the key is labeled A on a French keyboard.

By switching from keyCode to code, TinyPilot should be able to escape the work of creating manual mappings for every keyboard layout.

Fixes #354

* Add more cleanup

* Fix name of findKeyCode

* Ignore the location property of keyboard events

We no longer need it because we can check the orientation from the key code (e.g., ShiftLeft, MetaRight)

* Fix comment

* Simplify logic of detecting AltGraph key

* Remove unused key mappings

* Add test for too long key code

* Add paste support

* Remove old function reference

* Move shift key logic into keycodes.js

* Fix name of isKeyPressed

* Fix backtick processing

* Treat key code as a string

* Fix ordering of ignored keystroke check

* Remove redundant modifier definitions

* If a right modifier is pressed, assume it's the only one.

* Fix right modifier workaround

* Move isModifierCode to keycodes.js

* Handle AltGraph properly

* Fix variable declaration

* Fix key variable

* Fix detection of AltGraph keystroke

* Fix keycode handling

* Fix state management for AltGraph