Install ansible role requirements through requirements.yml (#382)
We're transitioning away from installing roles implicitly through meta/main.yml, as it doesn't allow us to apply any logic before the install of dependent roles. If we install via the requirements.yml, we can choose when in the role we want to pull in the parent roles and add more sophisticated logic to them instead of installing them right off the bat with the default variable assignments.