Make it possible to update TinyPilot through the UI
* New dropdown "Options -> Update" on navbar
* Two new custom elements update-{dialog,wait} which are both the
"update" and the "updating" dialogs
* A new update.js file that handles the functions related to updating
TinyPilot

The update button calls /version and /latestRelease to compare them and
see if a newer version is available (this is all based on commit IDs).

An "update-started" event is created when the user presses the "Update"
button. This will trigger the "update-wait" dialog showing a five minute
countdown timer. If there's no error during the process, a
reboot happens.