Return HTTP 200 for application-level errors
Previously, we were returning HTTP 500 for any application error, but that makes it harder to write frontend code that wants to read the 'error' field of the response.

All of the frontend REST code checks the success result anyway, so we can rely on that instead of the HTTP error code and reserve HTTP error codes for uncaught exceptions that really surprise the app.