Only do a 30s countdown during shutdown, not restart
This fixes a bug in the logic flow. Previously, both restart and shutdown kicked off a 30 second timer to announce a shutdown is complete. The correct behavior is for restart to spin forever until it detects the server's websocket connection has been reestablished, and the shutdown flow should assume the server is shut down after 30 seconds.