Make a fake success message shutdown returns 502 (#423)
There's a bug in the handling of the fetch request because it attempts to resolve the Promise as success, but the next stage rejects it for not having the success parameter explicitly set in the response body. This updates it to add a fake success message becaues 502 almost certainly means shutting down the server worked.