Change test file organization (#483)
Previously, tests were in a separate folder with a mirrored folder hierarchy. This made tests kind of hard to discover. Instead, the tests will now be alongside the production code and include a _test suffix.

The standard convention is to make it a test_ prefix, but making it a suffix means that in most file views, the test file is adjacent to the file it's testing, which is a substantial win.