Remove CORS workaround for Flask-SocketIO
I originally added wildcard support so that users could access TinyPilot from addressess like http://tinypilot or http://10.0.0.123 and it was breaking when CORS was hardcoded to a specific address. But the better fix is to just remove the CORS handling altogether and accept Flask-SocketIO's default behavior of same-origin, which will work regardless of whether the user accesses by IP, hostname, or even a remapped hostname that the server doesn't know about.