Adjust timing of update logic
The post-update check was treating the timestamp as the update completion time, but the update writer was only updating the timestamp at update start. As a result, TinyPilot would never see updates complete if they took longer than 3 minutes because 3 mins was the cutoff.

This relaxes the cutoff to 5 + 3 = 8 minutes and also adjusts the update logic so that it refreshes the timestamp at the end of the update.