Listen only on localhost by default (#615)
* Listen only on localhost by default

Listening on all interfaces is really only useful during development. In production, nginx forwards requests to TinyPilot, so TinyPilot doesn't need to listen on any non-localhost interface.

While I can't think of any security issues related to this, it's safer if we reduce the attack surface by listening only on on interfaces that are strictly required in production.

* Set the default listen interface to 127.0.0.1 explicitly

socketio.run defaults to this value, but we may as well be explicit.