Use HTTP instead of socket.io to detect when an update is complete
When we upgraded from socket.io protocol 2 to 3, it broke the check for whether a restart had completed because the frontend was still on protocol 2, and the backend had upgraded to protocol 3. We should rely on regular HTTP to check the status after an update so that we're not vulnerable to this sort of bug in the future.