Change semantics of update settings module
The update settings module had confusing naming and was awkward for callers to use because they had to directly open the settings file and pass it to the module. It wasn't so bad when there was just one caller, but it looks like we'll potentially have many callers, so it makes sense to make the calling semantics easier and more obbvious.

Fixes #650