Use error=None to represent update success (#659)
The previous implementation was inconsistent about what value the update status API returned for the error field. Through some code paths, it would return an empty string for the error field on success, and for other code paths, it returned null.

This changes the update modules so that they consistently use an error value of None/null to represent a successful update result.