Add --force to git fetch command
If the TinyPilot has any commits or tags that conflict with the server, the server's version should override the device's.

This has caused issues in cases where we accidentally published conflicting release tags, but then device that saw the old release tag error out if they see the new one.