Apply a uStreamer workaround for browser bugs in MJPEG handling
Applies uStreamer's dual_final_frames=1 option to work around a bug in how Chrome and other browsers handle MJPEG when uStreamer's --drop-same-frames option is in use.

uStreamer's documentation suggests that this is a Safari/Webkit-specific workaround, but I've verified that this affects Chrome, Firefox, and Edge as well.

Fix confirmed on:

* Chrome 93 x64 (Windows)
* Edge 94 x64 (Windows)
* Firefox 92 x64 (Windows)