Reimplement throttling logic for mouse events (#798)
* Reimplement throttling logic for mouse events

The previous mouse throttling logic used the underscore.js throttle() function. It worked well enough, but the implementation was ugly because we had several different hacks in place to work around library loading issues or handling changes to throttling settings mid-throttle.

The biggest problem was that we configured throttle() to drop events along the trailing edge of a throttle window. This had the benefit of reducing latency for the first mouse event but the drawback that it potentially drops later mouse events that are more important (as final mouse position is more important than intermediate positions).

This change reimplements the rate-limiting logic without a third-party library. Instead, we maintain a queue of mouse events capped at 10 events max. Every time we queue an event, we start a timer to process subsequent elements. Because we're setting the timer on queueing events, we can correctly update the rate limit time based on the latest value of milliseconds-between-mouse-events rather than resort to hacks that approximate it, like we did in the previous implementation.

Fixes #795

* Clarify comment

* Refactor mouse logic into mouse.js

* Document rate limited mouse behavior

* Add more comments for mouse down and mouse up priority

* onMouseWheel -> onWheel

* Add documentation and simplify implementation for single-event queue

* Update documentation

* Small refactoring

* Refactor

* Clarify comments

* Updating based on code review notes

* Fix bug in implementation

* Update based on code review notes