Upgrade to cryptography==35.0.0 in ansible environment (#810)
TinyPilot bootstraps an Ansible environment in order to install TinyPilot in quick-install. We've been stuck on an old version of the cryptography package for a while because it required a Rust environment to install for a while, but now binary builds are available in wheel, so we don't need Rust to build it anymore, and we may as well be on the latest version.

This also lets us get rid of the workaround where we had to explicitly install the wheel package first. Now we can upgrade to a recent version of pip and it will use binary wheels where available instead of building packages locally.