Fix JS to HID translation of numpad add and subtract
TinyPilot was failing to forward the numpad keys + and - (on the numpad, not on the number row for en-us keyboards). The correct JavaScript names for these keys are NumpadAdd and NumpadSubtract. I don't think this is a recent change; it's more likely that I just typed in the numbers in incorrectly in #355.