Move negation of mouse scroll wheel to the server
JS and HID have opposite semantics for representing vertical scrolling. JavaScript represents downward scrolling with a positive number, whereas HID indicates it with a negative number.

We were previously flipping the sign in JS, but that's a poor place for it because the server is responsible for translating between JS events and HID messages, so this change moves the translation to the server.