Set explicit format for log messages
This fixes a bug where TinyPilot's log formatter at some point in our history got disconnected from the root logger, so we were creating a log formatter and discarding it without applying it to the actual logger.

This applies our log format to the root logger and adds in milliseconds-level precision because the default Flask logger includes it, and it might be useful for debugging at some point.