Give clearer names to the Python websocket connection handlers (#879)
* Giving clearer names to the Python websocket connect and disconnect handlers

These names date back to the original addition of websockets to the app (b45c7ccf3437aafc10b19d6e5874911e5ccc00f1). I used the names test_connect and test_disconnect based on the Flask-SocketIO documentation, thinking that the test_ prefix was a convention. I now realize it was just for demonstration purposes, and the test_ prefix is meaningless.

https://flask-socketio.readthedocs.io/en/latest/getting_started.html#connection-events

This changes the names to on_connect and on_disconnect so the functions are not confused for test functions.

* Make all function names in socket_api.py consistent