Remove 'Copy to Clipboard' button on logs page
The Copy to Clipboard button is primarily so that users can get technical support, and there are no technical support channels where pasting the full logs directly into the communication is the right thing. If they paste it into email or our support forum, the formatting gets mangled and it's very hard to read.

It's better if users just do 'Get Shareable URL' and share the URL with us.

This simplifies the page to remove the unneeded 'Copy to Clipboard' button for copying full logs to the clipboard.

If the user *really* wants to copy everything to their clipboard, they still can by just highlighting the text and using the browser's normal text copy functionality.