Put template for About dialog credits in a separate file
Previously, we broke our typical pattern for custom elements by placing the template for entries in the list of credits was in a template within the same file as the about page dialog rather than in its own file, as we normally do.

Combining the element had two unintended I overlooked: it actually duplicates entries multiple times in the list, and it includes the same <style> tag and rules ~40 times in the HTML.

This change ungroups the custom elements so that the template for entries in the credits has its own distinct file and follows our normal patterns for custom elments.