Upgrade to pylint 2.14.2
Upgrades pylint in dev_requirements.txt to pylint 2.14.2 and adjusts .pylintrc to match changes in pylint's configuration.

This change makes a few trivial code changes to fix new pylint warnings, but it leaves the bulk of the new warnings as separate work items so that we can keep the scope small.