Upgrade pip to a minimum version rather than a specific version
We added a requirement on pip 21.3.1 when the standard pip that shipped with Raspberry Pi OS was older than 21.3.1, and we wanted a newer feature. Now, Raspberry Pi OS ships a newer version of pip and we're forcing the system to our particular version, which is unnecessary.

This change relaxes the requirement on the pip version to be 21.3.1 or later, so we only have to reinstall pip if the system's version is pretty old.