Prevent users from accidentally downgrading from TinyPilot Pro to Community
We have this check in our legacy quick-install script (https://github.com/tiny-pilot/tinypilot/blob/fa2a5fed814e3ae80415b58918d71e28fbcd10c7/quick-install#L95L132) but we haven't ported it over to our new flow.

We're changing it slightly because the previous implementation tried to accomodate systems where TinyPilot was installed to a path other than /opt/tinypilot. Too much of our system assumes that TinyPilot is in /opt/tinypilot, so it's not worth the complexity of accomodating a different install location, so we just hardcode the directory to /opt/tinypilot.