Use full window width to display remote screen
I spoke to a customer who pointed out that this is wasted space, especially when the window size is reduced, and we could just use the full width of the window.

I originally kept the margins just because it looked a little tidier not to have it extend completely into the edges, but testing the full bleed behavior does look about the same aesthetically, and it seems like an improvement pragmatically.