Fix recommended VS code extensions
Updating VS Code's extensions to match what's in the actual project. vetur is a Vue plugin, so we don't need that. eslint might be useful, but we aren't integrated with it yet in any of the other project config, so we should exclude it until we decide to use it. python, we're currently using in settings.json, but we forgot to include it as a recommended extension.