Refactor how we place files in Debian package
This change adjusts how we package files in the Debian package so that we include less by default.

When we originally discussed whether to err on the side of including too much or too little, we didn't consider that including everything makes it more difficult to express where to place files in the Debian package.

This change adds a debian-pkg folder to the root of the repo, and everything within that folder gets placed into the Debian package root. Right now, this allows us to extract the special Debian preinst, postinst, prerm scripts to actual files, which allows shellcheck to flag issues that it can't flag when they're embedded in a Dockerfile.

The other benefit is that it allows us to embed other files that we want to place on the device within this repo. I've created a rough proof-of-concept of how we can continue this pattern to include all of /opt/tinypilot-privileged's scripts in the tinypilot core repo instead of keeping them separated in the Ansible role: https://github.com/tiny-pilot/tinypilot-pro/pull/618