Refactor how we place files in Debian package (#1146)
* Refactor how we place files in Debian package

This change adjusts how we package files in the Debian package so that we include less by default.

When we originally discussed whether to err on the side of including too much or too little, we didn't consider that including everything makes it more difficult to express where to place files in the Debian package.

This change adds a debian-pkg folder to the root of the repo, and everything within that folder gets placed into the Debian package root. Right now, this allows us to extract the special Debian preinst, postinst, prerm scripts to actual files, which allows shellcheck to flag issues that it can't flag when they're embedded in a Dockerfile.

The other benefit is that it allows us to embed other files that we want to place on the device within this repo. I've created a rough proof-of-concept of how we can continue this pattern to include all of /opt/tinypilot-privileged's scripts in the tinypilot core repo instead of keeping them separated in the Ansible role: https://github.com/tiny-pilot/tinypilot-pro/pull/618

* DEBUG: print bundle contents

* Revert "DEBUG: print bundle contents"

This reverts commit 46d1a5daeaa75dd940efd96d81ead6c4ff3aae1d.

* Fix blank line and bash lint in prerm

* Move Dockerfile to debian-pkg

* Simplify dockerignore

* DEBUG: print bundle contents

This reverts commit b52913b910defa875b3ae572ea15e7a3a21cdbdb.

* Fix Dockerfile context

* DEBUG: print bundle contents

This reverts commit b52913b910defa875b3ae572ea15e7a3a21cdbdb.

* Ignore Dockerfile in building Debian package

* Fix path to Dockerfile

* Revert "DEBUG: print bundle contents"

This reverts commit fbcc632c7323ff98f4cd9ea68c6a988a4a10fbc4.