Transfer privileged scripts to core repo
Previously, TinyPilot scripts that required root privileges lived in the ansible-role-tinypilot repo. This was due to security considerations for installing TinyPilot via Ansible. Our security model assumed that the tinypilot user had write access to everything in the TinyPilot git repo, so we couldn't keep scripts with sudo privileges in a place where tinypilot could modify them, as that would effectively give tinypilot root privileges (if tinypilot can modify a script and run it as root, it effectively has arbitrary root permisssions).

When we switched to Debian-based installs, we no longer had to worry about the tinypilot user being able to access files in the TinyPilot git repo, as the Debian package manager now provisions the device with the proper files and puts them in the right place.

This has a few benefits:
* If we need to call a privileged script from the TinyPilot web app, we no longer have to coordinate changes across two repos
* Installing privilged scripts with Debian instead of Ansible is significantly faster
* The Debian package manager takes responsibility for cleaning up unused scripts on upgrade/uninstall