Build Debian package with simple Docker rather than Docker+QEMU
I'm not sure why I thought we needed to use QEMU to build the Debian package, as there are no architecture-specific files in the package. It might have just been that we copy/pasted from Janus without thinking about it too much.

This simplifies the Debian package build to just use standard Docker so that it runs faster and can build on more systems (e.g., those that don't support QEMU).