Move Docker args below apt-get steps
Normall, I like having all the args at the beginning, but in this case, it makes debugging harder because it prevents Docker from caching the layer that does apt-get update and install, which is a bottleneck in our build process.

This won't make a difference in CI because it only runs the build once there, but it drastically reduces Debian package build time during development.