Fix Debian package filename to use arch all instead of armhf
We have a mismatch between the architecture we declare in the Debian package and the architecture we specify in the filename. The filename was specifying an architecture of armhf, while the Debian package was declaring all. all is correct, so this change updates the filename to match and makes it harder for the two to go out of sync.