Move Docker args below apt-get steps (#1195)
* Move Docker args below apt-get steps

Normall, I like having all the args at the beginning, but in this case, it makes debugging harder because it prevents Docker from caching the layer that does apt-get update and install, which is a bottleneck in our build process.

This won't make a difference in CI because it only runs the build once there, but it drastically reduces Debian package build time during development.

* Fix Debian package filename to use arch all instead of armhf (#1197)

We have a mismatch between the architecture we declare in the Debian package and the architecture we specify in the filename. The filename was specifying an architecture of armhf, while the Debian package was declaring all. all is correct, so this change updates the filename to match and makes it harder for the two to go out of sync.