Build TinyPilot Debian package and install bundle outside the main branch (#1214)
This change adjusts filtering in CircleCI so that we build the TinyPilot
Debian package and install bundle even in non-main repo branches.

Uploading a bundle changes state on our update server, so we should only
do that in our main branch. But all the other steps are fine to run in
branches because it doesn't affect state outside of CI. It's often
useful to build packages or bundles in the branches if we want to test
something end-to-end.