Verify working directory in create-bundle
create-bundle implicitly assumes that the current working directory is the ./bundler directory. This is error-prone, as our other TinyPilot scripts that depend on working directories set the directory themselves.

This change makes create-bundle take responsibility for setting the working directory properly so that callers can execute the script regardless of external working directory.