Fix paths to third-party licenses in Python 3.9 environments
On Buster, the default Python version is 3.7, so our paths referenced that Python version. On Bullseye, the default Python version is 3.9, which broke our paths to third-party licenses.

This change changes the version to 3.* so that it will match any Python version. There's a risk that we'll match the wrong version package, but I think that's less likely than accidentally breaking links to licenses.