Extract prompt part from `<update-dialog>` component (#1315)
This PRs is part of the proper implementation of the [POC for the update
flow](https://github.com/tiny-pilot/tinypilot-pro/pull/764). It lays the
groundwork for us to resolve
https://github.com/tiny-pilot/tinypilot/issues/1284 in Pro.

This is a non-functional refactoring. A few notes, in addition to what
we discussed in the POC:

- I took over your suggestion to place the prompt component right from
the start, and I like that much better than what I initially came up
with.
- I eventually decided to pull out the prompt component instead of
keeping the different prompts inside `<update-dialog>` next to each
other. The most compelling reason for me was that this will allow us to
minimize the differences between Pro and Community in the
`update-dialog.html` file. In Pro, we later only have to add a new,
separate file for the `<update-prompt-manual>` component, and do minimal
additional wiring in `<update-dialog>`. The downside is that we pay for
that separation with some boilerplate code (which I’d mind less here,
though).
- In the frontend, I slightly refactored the `getVersion` controller
function, to make the return structure on par with the one from
`getLatestRelease`. That’s for purely aesthetic reasons.

Note to self: when doing the upstream merge into Pro, we need to take
over the changelog link, which is Pro-specific.

(Extracted from https://github.com/tiny-pilot/tinypilot/pull/1312.)
<a data-ca-tag
href="https://codeapprove.com/pr/tiny-pilot/tinypilot/1315"><img
src="https://codeapprove.com/external/github-tag-allbg.png" alt="Review
on CodeApprove" /></a>