Introduce standalone `<share-logs-button>` component (#1335)
Part of https://github.com/tiny-pilot/tinypilot/issues/1307.

This PR introduces a new frontend component `<share-logs-button>`, that
allows us to re-use the logs sharing functionality in other contexts as
well, e.g. for sharing error logs (from the `<error-dialog>`).

The look and usage is demonstrated in
https://github.com/tiny-pilot/tinypilot/pull/1336.

A few notes:

- I expanded the comment in `clipboard.js` a bit. I also realized that
we might have to replace this workaround, as `document.execCommand` is
deprecated. For now, we should still be good, though.
- I parametrized the colors for the `<progress-spinner>`, to adjust the
contrast to fit the disabled “action” button. I’m not sure we have
enough of a universal pattern yet to document something like an
“in-progress” button in the style guide, since this is the only instance
so far.
- I was debating how we pass the logs text into this component:
currently, the `initialize()` method takes as supplier callback; but of
course, we could also pass in the logs text as string copy directly. I
don’t feel super strongly here, but it felt better to me to not make a
hard copy at initialization time, but to rather “freshly request” the
logs text when the user actually presses the button. The effect is the
same for both approaches, so this is mostly a stylistic issue, and maybe
also a question of data ownership.
- Currently, when `logs.tinypilotkvm.com` would fail for whatever
reason, the button would just bounce back to its initial state. If we
feel that we should give clearer feedback about this scenario, we could
consider e.g. toggling the button label to “Error!” for 2 seconds. (That
would have some extra complexity, though, so I thought it might be fine
as is, assuming that the failure is not too likely.)
<a data-ca-tag
href="https://codeapprove.com/pr/tiny-pilot/tinypilot/1335"><img
src="https://codeapprove.com/external/github-tag-allbg.png" alt="Review
on CodeApprove" /></a>