Use playwright to test navigation to about page, dependencies pages, and licenses (#832)
The manual end-to-end test process impedes the release cycle. So, we've
decided to experiment with semi-automating the manual testing of the UI.
We want to test how well playwright can replace manually testing the UI
in browsers with a command-line execution of the end-to-end tests.

As the first example of a Playwright test and to see how Playwright can
integrate into the release process, we test navigating to the About page
as well as the TinyPilot Privacy Policy and License, The dependencies'
pages and dependencies' licenses.

Fixes #836

You can install playwright with

```
npm install && \
sudo npx playwright install-deps
```

You can run run this test with

```
./dev-scripts/run-e2e-tests
```

You can step through the test in the [Playwright
Inspector](https://playwright.dev/docs/debug#playwright-inspector) with

```
PWDEBUG=1 ./dev-scripts/run-e2e-tests --ui
```

* To cut scope, I only wrote a test of the About dialogue. I explored
testing the debug logs and discovered that it requires both interaction
with the clipboard and causes the dev server to ask the user to enter a
password. It is wiser to tackle in a separate task after getting
feedback.

* To cut scope, I start out running the tests to run against the dev
server running at http://0.0.0.0:8000 instead of https://tinypilot.

* To give myself time to see *how* I want to simplify the test code,
leave comments about what to improve about code quality. Usually we
[value quality and maintainability over turnaround
time](https://mtlynch.io/freelancer-guidelines/#code-quality) but [test
code is not like other
code](https://mtlynch.io/good-developers-bad-tests/#test-code-is-not-like-other-code)
and I want to give myself time to see how I will make the test code more
obvious.

---------

Co-authored-by: Andrew M. Farrell <armorsmith42@gmail.com>